Avoid the Haskell conformance tests depending on plutus-metatheory
